data_loading_scripts/pyproject.toml
[project]
name = "mealdb-data-loader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "supabase>=2.28.3",
]
